Move census domain into census app
