+++ accounts/schema/mutation.py
@@ -28,13 +28,6 @@ from accounts.schema.mutations import (
     AdminVillageCapabilityUpdateMutation,
     AdminVillageCreateMutation,
     AdminVillageUpdateMutation,
-    AdminAnimalCensusCapabilityUpdateMutation,
-    AdminAnimalSpeciesCreateMutation,
-    AdminAnimalSpeciesUpdateMutation,
-    AdminCensusDefinitionsEnsureDefaultsMutation,
-    AdminCensusDefinitionVersionPublishMutation,
-    SubmitVillageCensusSnapshotMutation,
-    SubmitVillageCensusSnapshotV2Mutation,
     RequestToDeleteMyAccountMutation,
 )
 
@@ -69,17 +62,4 @@ class Mutation(graphene.ObjectType):
     admin_village_capability_update = AdminVillageCapabilityUpdateMutation.Field()
     admin_village_create = AdminVillageCreateMutation.Field()
     admin_village_update = AdminVillageUpdateMutation.Field()
-    admin_animal_census_capability_update = (
-        AdminAnimalCensusCapabilityUpdateMutation.Field()
-    )
-    admin_animal_species_create = AdminAnimalSpeciesCreateMutation.Field()
-    admin_animal_species_update = AdminAnimalSpeciesUpdateMutation.Field()
-    admin_census_definitions_ensure_defaults = (
-        AdminCensusDefinitionsEnsureDefaultsMutation.Field()
-    )
-    admin_census_definition_version_publish = (
-        AdminCensusDefinitionVersionPublishMutation.Field()
-    )
-    submit_village_census_snapshot = SubmitVillageCensusSnapshotMutation.Field()
-    submit_village_census_snapshot_v2 = SubmitVillageCensusSnapshotV2Mutation.Field()
     request_to_delete_my_account = RequestToDeleteMyAccountMutation.Field()
