+++ census/animal_census_capability.py
