+++ census/schema/__init__.py
@@ -1,0 +1,2 @@
+from .query import Query
+from .mutation import Mutation
