+++ census/schema/mutations/__init__.py
@@ -1,0 +1,4 @@
+from .admin_animal_census_capability_mutations import *
+from .admin_animal_species_mutations import *
+from .admin_census_definition_mutations import *
+from .village_census_mutations import *
