+++ podd_api/settings.py
@@ -74,6 +74,7 @@ SHARED_APPS = (
     "channels",
     "common",
     "reports",
+    "census",
     "cases",
     "notifications",
     "summaries",
@@ -89,6 +90,7 @@ TENANT_APPS = (
     "common",
     "accounts",
     "reports",
+    "census",
     "cases",
     "notifications",
     "summaries",
